:truck: renamed Meldung to Praktikumsplaetze
