+++ praktikumsplaner-frontend/frontend/src/views/Praktikumsplaetze/MeldungStudium.vue
