+++ praktikumsplaner-frontend/frontend/src/views/Praktikumsplaetze/PraktikumsplaetzeView.vue
